requirements.txt
pycryptodome
gmpy2
sympy
pwntools
numpy
matplotlib